Refactor view gateway and runtime assembly for modularity
