+++ adapters/telegram/gateway/sender.py
@@ -1,0 +1,63 @@
+"""Message sending helpers for the Telegram gateway."""
+from __future__ import annotations
+
+from aiogram import Bot
+
+from navigator.core.port.limits import Limits
+from navigator.core.port.markup import MarkupCodec
+from navigator.core.port.message import Result
+from navigator.core.port.extraschema import ExtraSchema
+from navigator.core.port.pathpolicy import MediaPathPolicy
+from navigator.core.port.preview import LinkPreviewCodec
+from navigator.core.telemetry import Telemetry, TelemetryChannel
+from navigator.core.value.content import Payload
+from navigator.core.value.message import Scope
+
+from .send import SendContextFactory, SendDispatcher
+from ..serializer.screen import SignatureScreen
+
+
+class TelegramMessageSender:
+    """Coordinate Telegram send operations for navigator payloads."""
+
+    def __init__(
+        self,
+        bot: Bot,
+        *,
+        codec: MarkupCodec,
+        limits: Limits,
+        schema: ExtraSchema,
+        policy: MediaPathPolicy,
+        screen: SignatureScreen,
+        preview: LinkPreviewCodec | None,
+        truncate: bool,
+        telemetry: Telemetry,
+    ) -> None:
+        self._dispatcher = SendDispatcher(
+            bot,
+            schema=schema,
+            screen=screen,
+            policy=policy,
+            limits=limits,
+            telemetry=telemetry,
+        )
+        self._context_factory = SendContextFactory(codec=codec, preview=preview)
+        self._truncate = truncate
+        self._channel: TelemetryChannel = telemetry.channel(__name__)
+
+    async def send(self, scope: Scope, payload: Payload) -> Result:
+        context = self._context_factory.build(
+            scope=scope,
+            payload=payload,
+            channel=self._channel,
+        )
+        message, extras, meta = await self._dispatcher.dispatch(
+            payload,
+            scope=scope,
+            context=context,
+            truncate=self._truncate,
+        )
+        return Result(id=message.message_id, extra=extras, meta=meta)
+
+
+__all__ = ["TelegramMessageSender"]
